refactor: Extracted function (hasRegistry)
