+++ src/steps/create-registries.ts
@@ -11,31 +11,11 @@ import {
   type TransformedEntityName,
   transformEntityName,
 } from '../utils/files.js';
+import { hasRegistry } from './create-registries/index.js';
 
 type Data = {
   entity: TransformedEntityName;
 };
-
-function hasRegistry(file: string): boolean {
-  const traverse = AST.traverse(true);
-
-  let hasRegistry = false;
-
-  traverse(file, {
-    visitTSModuleDeclaration(path) {
-      // @ts-ignore: Assume that types from external packages are correct
-      const moduleName = path.node.id.value;
-
-      if (moduleName === '@glint/environment-ember-loose/registry') {
-        hasRegistry = true;
-      }
-
-      return false;
-    },
-  });
-
-  return hasRegistry;
-}
 
 function createRegistry(file: string, data: Data): string {
   const traverse = AST.traverse(true);
