+++ src/steps/create-registries/index.ts
@@ -1,0 +1,1 @@
+export * from './has-registry.js';
